Started documenting and moving into correct packages, invoice body classes
