+++ src/main/java/com/github/siboxd/fatturapa/model/DettaglioLineeType.java
@@ -1,5 +1,8 @@
 package com.github.siboxd.fatturapa.model;
 
+import com.github.siboxd.fatturapa.model.invoicebody.general.pensionfund.NaturaType;
+import com.github.siboxd.fatturapa.model.invoicebody.general.pensionfund.RitenutaType;
+
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.ElementList;
 import org.simpleframework.xml.Root;
