--- src/main/java/com/github/siboxd/fatturapa/model/TipoCassaType.java
@@ -1,33 +1,0 @@
-package com.github.siboxd.fatturapa.model;
-
-import org.simpleframework.xml.Namespace;
-import org.simpleframework.xml.Root;
-
-@Root(name = "TipoCassaType")
-@Namespace(reference = "http://ivaservizi.agenziaentrate.gov.it/docs/xsd/fatture/v1.2")
-public enum TipoCassaType {
-
-    TC01,
-    TC02,
-    TC03,
-    TC04,
-    TC05,
-    TC06,
-    TC07,
-    TC08,
-    TC09,
-    TC10,
-    TC11,
-    TC12,
-    TC13,
-    TC14,
-    TC15,
-    TC16,
-    TC17,
-    TC18,
-    TC19,
-    TC20,
-    TC21,
-    TC22
-
-}
