--- src/main/java/com/github/siboxd/fatturapa/model/TipoRitenutaType.java
@@ -1,13 +1,0 @@
-package com.github.siboxd.fatturapa.model;
-
-import org.simpleframework.xml.Namespace;
-import org.simpleframework.xml.Root;
-
-@Root(name = "TipoRitenutaType")
-@Namespace(reference = "http://ivaservizi.agenziaentrate.gov.it/docs/xsd/fatture/v1.2")
-public enum TipoRitenutaType {
-
-    RT01,
-    RT02
-
-}
