+++ src/main/java/com/github/siboxd/fatturapa/model/invoicebody/FatturaElettronicaBodyType.java
@@ -1,4 +1,10 @@
-package com.github.siboxd.fatturapa.model;
+package com.github.siboxd.fatturapa.model.invoicebody;
+
+import com.github.siboxd.fatturapa.model.AllegatiType;
+import com.github.siboxd.fatturapa.model.DatiBeniServiziType;
+import com.github.siboxd.fatturapa.model.DatiPagamentoType;
+import com.github.siboxd.fatturapa.model.DatiVeicoliType;
+import com.github.siboxd.fatturapa.model.invoicebody.general.DatiGeneraliType;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.ElementList;
@@ -8,7 +14,10 @@ import java.util.List;
 
 
 /**
- * FatturaElettronicaBodyType<br>
+ * This block - which constitutes the body of the invoice and represents its specific content -
+ * must be present only once in the case of a single invoice, whereas in the case of a batch
+ * of invoices the lot itself must be repeated for each invoice.
+ * <p>
  * Generated using Android JAXB<br>
  *
  * @link https://github.com/yeshodhan/android-jaxb
@@ -16,14 +25,18 @@ import java.util.List;
 @Root(name = "FatturaElettronicaBodyType")
 public class FatturaElettronicaBodyType {
 
-    @Element(name = "DatiGenerali", required = true)
+    @Element(name = "DatiGenerali")
     private DatiGeneraliType datiGenerali;
-    @Element(name = "DatiBeniServizi", required = true)
+
+    @Element(name = "DatiBeniServizi")
     private DatiBeniServiziType datiBeniServizi;
+
     @Element(name = "DatiVeicoli", required = false)
     private DatiVeicoliType datiVeicoli;
+
     @ElementList(name = "DatiPagamento", entry = "DatiPagamento", inline = true, required = false)
     private List<DatiPagamentoType> datiPagamento;
+
     @ElementList(name = "Allegati", entry = "Allegati", inline = true, required = false)
     private List<AllegatiType> allegati;
 
@@ -34,7 +47,7 @@ public class FatturaElettronicaBodyType {
         return datiGenerali;
     }
 
-    public void setDatiGenerali(DatiGeneraliType datiGenerali) {
+    public void setDatiGenerali(final DatiGeneraliType datiGenerali) {
         this.datiGenerali = datiGenerali;
     }
 
@@ -42,7 +55,7 @@ public class FatturaElettronicaBodyType {
         return datiBeniServizi;
     }
 
-    public void setDatiBeniServizi(DatiBeniServiziType datiBeniServizi) {
+    public void setDatiBeniServizi(final DatiBeniServiziType datiBeniServizi) {
         this.datiBeniServizi = datiBeniServizi;
     }
 
@@ -50,7 +63,7 @@ public class FatturaElettronicaBodyType {
         return datiVeicoli;
     }
 
-    public void setDatiVeicoli(DatiVeicoliType datiVeicoli) {
+    public void setDatiVeicoli(final DatiVeicoliType datiVeicoli) {
         this.datiVeicoli = datiVeicoli;
     }
 
@@ -58,7 +71,7 @@ public class FatturaElettronicaBodyType {
         return datiPagamento;
     }
 
-    public void setDatiPagamento(List<DatiPagamentoType> datiPagamento) {
+    public void setDatiPagamento(final List<DatiPagamentoType> datiPagamento) {
         this.datiPagamento = datiPagamento;
     }
 
@@ -66,7 +79,7 @@ public class FatturaElettronicaBodyType {
         return allegati;
     }
 
-    public void setAllegati(List<AllegatiType> allegati) {
+    public void setAllegati(final List<AllegatiType> allegati) {
         this.allegati = allegati;
     }
