+++ src/main/java/com/github/siboxd/fatturapa/model/invoicebody/general/CausalePagamentoType.java
@@ -1,8 +1,13 @@
-package com.github.siboxd.fatturapa.model;
+package com.github.siboxd.fatturapa.model.invoicebody.general;
 
 import org.simpleframework.xml.Namespace;
 import org.simpleframework.xml.Root;
 
+/**
+ * It serves to highlight the reason for payment.
+ * <br><br>
+ * These values are the coding of the reason for payment on the <em>Model 770S</em>.
+ */
 @Root(name = "CausalePagamentoType")
 @Namespace(reference = "http://ivaservizi.agenziaentrate.gov.it/docs/xsd/fatture/v1.2")
 public enum CausalePagamentoType {
