+++ src/main/java/com/github/siboxd/fatturapa/model/invoicebody/general/DatiBolloType.java
@@ -1,11 +1,12 @@
-package com.github.siboxd.fatturapa.model;
+package com.github.siboxd.fatturapa.model.invoicebody.general;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
 
 
 /**
- * DatiBolloType<br>
+ * Describes stamp duty
+ * <p>
  * Generated using Android JAXB<br>
  *
  * @link https://github.com/yeshodhan/android-jaxb
@@ -13,9 +14,10 @@ import org.simpleframework.xml.Root;
 @Root(name = "DatiBolloType")
 public class DatiBolloType {
 
-    @Element(name = "BolloVirtuale", required = true)
+    @Element(name = "BolloVirtuale")
     private BolloVirtualeType bolloVirtuale;
-    @Element(name = "ImportoBollo", required = true)
+
+    @Element(name = "ImportoBollo")
     private String importoBollo;
 
     public DatiBolloType() {
@@ -25,7 +27,7 @@ public class DatiBolloType {
         return bolloVirtuale;
     }
 
-    public void setBolloVirtuale(BolloVirtualeType bolloVirtuale) {
+    public void setBolloVirtuale(final BolloVirtualeType bolloVirtuale) {
         this.bolloVirtuale = bolloVirtuale;
     }
 
@@ -33,7 +35,12 @@ public class DatiBolloType {
         return importoBollo;
     }
 
-    public void setImportoBollo(String importoBollo) {
+    /**
+     * @param importoBollo The field contains a numeric value consisting of an integer and two
+     *                     decimal places. The decimals, separated by the whole with the dot character
+     *                     ("."), Must always be indicated even if zero (eg: 25.00).
+     */
+    public void setImportoBollo(final String importoBollo) {
         this.importoBollo = importoBollo;
     }
