+++ src/main/java/com/github/siboxd/fatturapa/model/invoicebody/general/DatiGeneraliDocumentoType.java
@@ -1,4 +1,8 @@
-package com.github.siboxd.fatturapa.model;
+package com.github.siboxd.fatturapa.model.invoicebody.general;
+
+import com.github.siboxd.fatturapa.model.Art73Type;
+import com.github.siboxd.fatturapa.model.ScontoMaggiorazioneType;
+import com.github.siboxd.fatturapa.model.invoicebody.general.pensionfund.DatiCassaPrevidenzialeType;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.ElementList;
@@ -8,7 +12,8 @@ import java.util.List;
 
 
 /**
- * DatiGeneraliDocumentoType<br>
+ * This block contains the general data of the invoice document.
+ * <p>
  * Generated using Android JAXB<br>
  *
  * @link https://github.com/yeshodhan/android-jaxb
@@ -16,28 +21,39 @@ import java.util.List;
 @Root(name = "DatiGeneraliDocumentoType")
 public class DatiGeneraliDocumentoType {
 
-    @Element(name = "TipoDocumento", required = true)
+    @Element(name = "TipoDocumento")
     private TipoDocumentoType tipoDocumento;
-    @Element(name = "Divisa", required = true)
+
+    @Element(name = "Divisa")
     private String divisa;
-    @Element(name = "Data", required = true)
+
+    @Element(name = "Data")
     private String data;
-    @Element(name = "Numero", required = true)
+
+    @Element(name = "Numero")
     private String numero;
+
     @Element(name = "DatiRitenuta", required = false)
     private DatiRitenutaType datiRitenuta;
+
     @Element(name = "DatiBollo", required = false)
     private DatiBolloType datiBollo;
+
     @ElementList(name = "DatiCassaPrevidenziale", entry = "DatiCassaPrevidenziale", inline = true, required = false)
     private List<DatiCassaPrevidenzialeType> datiCassaPrevidenziale;
+
     @ElementList(name = "ScontoMaggiorazione", entry = "ScontoMaggiorazione", inline = true, required = false)
     private List<ScontoMaggiorazioneType> scontoMaggiorazione;
+
     @Element(name = "ImportoTotaleDocumento", required = false)
     private String importoTotaleDocumento;
+
     @Element(name = "Arrotondamento", required = false)
     private String arrotondamento;
+
     @ElementList(name = "Causale", entry = "Causale", inline = true, required = false)
     private List<String> causale;
+
     @Element(name = "Art73", required = false)
     private Art73Type art73;
 
@@ -48,7 +64,7 @@ public class DatiGeneraliDocumentoType {
         return tipoDocumento;
     }
 
-    public void setTipoDocumento(TipoDocumentoType tipoDocumento) {
+    public void setTipoDocumento(final TipoDocumentoType tipoDocumento) {
         this.tipoDocumento = tipoDocumento;
     }
 
@@ -56,31 +72,50 @@ public class DatiGeneraliDocumentoType {
         return divisa;
     }
 
-    public void setDivisa(String divisa) {
+    /**
+     * It is used to identify the account currency.
+     *
+     * @param divisa The field must contain the currency code used, according to
+     *               <em>ISO 4217 alpha-3: 2001</em> (eg "EUR" in the case of euros).
+     */
+    public void setDivisa(final String divisa) {
         this.divisa = divisa;
-    }
+    } // TODO: 20/01/2019 can be made checks??
 
     public String getData() {
         return data;
     }
 
-    public void setData(String data) {
+    /**
+     * It is a data required by the legislation and identifies the date of issue of the document.
+     *
+     * @param data The field contains the date of issue of the document in the format YYYY-MM-DD
+     *             (ISO 8601: 2004 standard).
+     */
+    public void setData(final String data) {
         this.data = data;
-    }
+    } // TODO: 20/01/2019 add checks
 
     public String getNumero() {
         return numero;
     }
 
-    public void setNumero(String numero) {
+    /**
+     * @param numero The field contains the document number and allows entry of alphanumeric values.
+     *               It must contain at least one numeric value (0-9).
+     */
+    public void setNumero(final String numero) {
         this.numero = numero;
-    }
+    } // TODO: 20/01/2019 add checks
 
     public DatiRitenutaType getDatiRitenuta() {
         return datiRitenuta;
     }
 
-    public void setDatiRitenuta(DatiRitenutaType datiRitenuta) {
+    /**
+     * <b>Note:</b> To be valued only if the seller is subject to withholding tax or on a permanent basis
+     */
+    public void setDatiRitenuta(final DatiRitenutaType datiRitenuta) {
         this.datiRitenuta = datiRitenuta;
     }
 
@@ -88,7 +123,10 @@ public class DatiGeneraliDocumentoType {
         return datiBollo;
     }
 
-    public void setDatiBollo(DatiBolloType datiBollo) {
+    /**
+     * <b>Note:</b> To be valued only if the stamp duty is required on the type of document/transaction.
+     */
+    public void setDatiBollo(final DatiBolloType datiBollo) {
         this.datiBollo = datiBollo;
     }
 
@@ -96,7 +134,12 @@ public class DatiGeneraliDocumentoType {
         return datiCassaPrevidenziale;
     }
 
-    public void setDatiCassaPrevidenziale(List<DatiCassaPrevidenzialeType> datiCassaPrevidenziale) {
+    /**
+     * <b>Note:</b> To be valued only if the seller is a party required to pay the social security
+     * contribution to their category cash desk or to the INPS (National Social Security Institute)
+     * (or both)
+     */
+    public void setDatiCassaPrevidenziale(final List<DatiCassaPrevidenzialeType> datiCassaPrevidenziale) {
         this.datiCassaPrevidenziale = datiCassaPrevidenziale;
     }
 
@@ -104,7 +147,7 @@ public class DatiGeneraliDocumentoType {
         return scontoMaggiorazione;
     }
 
-    public void setScontoMaggiorazione(List<ScontoMaggiorazioneType> scontoMaggiorazione) {
+    public void setScontoMaggiorazione(final List<ScontoMaggiorazioneType> scontoMaggiorazione) {
         this.scontoMaggiorazione = scontoMaggiorazione;
     }
 
@@ -112,7 +155,7 @@ public class DatiGeneraliDocumentoType {
         return importoTotaleDocumento;
     }
 
-    public void setImportoTotaleDocumento(String importoTotaleDocumento) {
+    public void setImportoTotaleDocumento(final String importoTotaleDocumento) {
         this.importoTotaleDocumento = importoTotaleDocumento;
     }
 
@@ -120,7 +163,7 @@ public class DatiGeneraliDocumentoType {
         return arrotondamento;
     }
 
-    public void setArrotondamento(String arrotondamento) {
+    public void setArrotondamento(final String arrotondamento) {
         this.arrotondamento = arrotondamento;
     }
 
@@ -128,7 +171,7 @@ public class DatiGeneraliDocumentoType {
         return causale;
     }
 
-    public void setCausale(List<String> causale) {
+    public void setCausale(final List<String> causale) {
         this.causale = causale;
     }
 
@@ -136,7 +179,7 @@ public class DatiGeneraliDocumentoType {
         return art73;
     }
 
-    public void setArt73(Art73Type art73) {
+    public void setArt73(final Art73Type art73) {
         this.art73 = art73;
     }
