+++ src/main/java/com/github/siboxd/fatturapa/model/invoicebody/general/TipoRitenutaType.java
@@ -1,0 +1,21 @@
+package com.github.siboxd.fatturapa.model.invoicebody.general;
+
+import org.simpleframework.xml.Namespace;
+import org.simpleframework.xml.Root;
+
+/**
+ * It is used to identify the type of withholding indicated on the invoice.
+ */
+@Root(name = "TipoRitenutaType")
+@Namespace(reference = "http://ivaservizi.agenziaentrate.gov.it/docs/xsd/fatture/v1.2")
+public enum TipoRitenutaType {
+    /**
+     * Withholding tax on natural persons
+     */
+    RT01,
+    /**
+     * Withholding tax on legal entities
+     */
+    RT02
+
+}
