+++ src/main/java/com/github/siboxd/fatturapa/model/invoicebody/general/pensionfund/NaturaType.java
@@ -1,0 +1,44 @@
+package com.github.siboxd.fatturapa.model.invoicebody.general.pensionfund;
+
+import org.simpleframework.xml.Namespace;
+import org.simpleframework.xml.Root;
+
+/**
+ * It serves to highlight the nature of the operation for which VAT is not required.
+ */
+@Root(name = "NaturaType")
+@Namespace(reference = "http://ivaservizi.agenziaentrate.gov.it/docs/xsd/fatture/v1.2")
+public enum NaturaType {
+    /**
+     * Operations Excluded with reference to Article 15 of Presidential Decree 633/1972
+     */
+    N1,
+    /**
+     * Operations Not subject
+     */
+    N2,
+    /**
+     * Non-Taxable Operations
+     */
+    N3,
+    /**
+     * Exempt Operations
+     */
+    N4,
+    /**
+     * Margin scheme
+     */
+    N5,
+    /**
+     * Reverse charge
+     */
+    N6,
+    /**
+     * VAT paid in other EU countries (distance sales pursuant to Article 40 paragraphs 3 and 4
+     * and Article 41 paragraph 1 letter b, Decree 331/93, provision of telecommunications,
+     * tele-broadcasting and electronic services pursuant to Article 7-sexies letter f, g,
+     * presidential decree 633/72 and article 74 sexies, presidential decree 633/72)
+     */
+    N7
+
+}
